Integrate Document into Index
